pyproject.toml
[project]
name = "lookae-spider"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=25.1.0",
    "aiologger>=0.7.0",
    "beautifulsoup4>=4.14.2",
    "bs4>=0.0.2",
    "chromadb>=1.3.0",
    "faiss-cpu>=1.12.0",
    "fastembed>=0.7.3",
    "flagembedding>=1.3.5",
    "huggingface-hub>=0.36.0",
    "jieba>=0.42.1",
    "justllms>=2.1.8",
    "langchain>=1.0.3",
    "langchain-anthropic>=0.3.22",
    "langchain-chroma>=0.2.6",
    "langchain-community>=0.4.1",
    "langchain-core>=1.0.2",
    "langchain-huggingface>=0.3.1",
    "langchain-openai>=1.0.1",
    "langchain-text-splitters>=0.3.11",
    "litellm>=1.79.0",
    "markdownify>=1.2.0",
    "openai>=2.6.1",
    "pypdf>=6.1.3",
    "python-dotenv>=1.2.1",
    "python-slugify>=8.0.4",
    "rank-bm25>=0.2.2",
    "requests>=2.32.5",
    "rich>=14.2.0",
    "sentence-transformers>=5.1.2",
    "tenacity>=9.1.2",
    "tqdm>=4.67.1",
]
